@@ -6,63 +6,61 @@
   "scripts": {
     "ng-high-memory": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng",
     "ng": "ng",
-    "start": "export NODE_OPTIONS=--openssl-legacy-provider; ng serve --proxy-config proxy-gxa.conf.json",
+    "start": "ng serve --proxy-config proxy-gxa.conf.json",
     "start-with-proxy": "ng serve --proxy-config proxy.conf.json",
-    "publish-gh-pages": "rm -rf dist; node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --output-path=dist --configuration production --base-href \"https://complex-portal.github.io/complex-portal-view/\"; sudo ngh -r https://github.com/Complex-Portal/complex-portal-view",
-    "publish-test-gh-pages": "rm -rf dist; node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --output-path=dist --configuration test --base-href \"https://complex-portal.github.io/complex-portal-view/\"; sudo ngh -r https://github.com/Complex-Portal/complex-portal-view",
-    "build-prod": "rm -rf dist; node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --configuration production --base-href \"/complexportal/\"",
+    "publish-gh-pages": "rm -rf dist; ng build --configuration production --base-href \"https://complex-portal.github.io/complex-portal-view/\"; ngh -r https://github.com/Complex-Portal/complex-portal-view",
+    "publish-test-gh-pages": "rm -rf dist; ng build --configuration test --base-href \"https://complex-portal.github.io/complex-portal-view/\"; ngh -r https://github.com/Complex-Portal/complex-portal-view",
+    "build-prod": "rm -rf dist; ng build --configuration production --base-href \"/complexportal/\"",
     "test": "ng test",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
     "e2e": "protractor"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.12",
-    "@angular/cdk": "^12.2.12",
-    "@angular/common": "^12.2.12",
-    "@angular/compiler": "^12.2.12",
-    "@angular/core": "^12.2.12",
-    "@angular/forms": "^12.2.12",
-    "@angular/material": "^12.2.12",
-    "@angular/platform-browser": "^12.2.12",
-    "@angular/platform-browser-dynamic": "^12.2.12",
-    "@angular/router": "^12.2.12",
+    "@angular/animations": "^18.1.4",
+    "@angular/cdk": "^18.1.4",
+    "@angular/common": "^18.1.4",
+    "@angular/compiler": "^18.1.4",
+    "@angular/core": "^18.1.4",
+    "@angular/forms": "^18.1.4",
+    "@angular/material": "^18.1.4",
+    "@angular/platform-browser": "^18.1.4",
+    "@angular/platform-browser-dynamic": "^18.1.4",
+    "@angular/router": "^18.1.4",
     "@ebi-gene-expression-group/anatomogram": "^2.4.0",
     "@swissprot/rhea-reaction-visualizer": "0.0.23",
-    "@types/estree": "^1.0.0",
     "angular-cli-ghpages": "^0.5.2",
     "animate.css": "^3.5.2",
-    "classlist.js": "^1.1.20150312",
-    "complexviewer": "^2.2.2",
+    "complexviewer": "^2.2.4",
     "core-js": "^2.5.3",
     "ebi-framework": "1.3.8",
     "expose-loader": "^5.0.0",
-    "jquery": "^3.4.1",
+    "jquery": "^3.7.1",
     "litemol": "https://github.com/dsehnal/LiteMol/tarball/master",
-    "ngx-google-analytics": "^11.2.1",
-    "ngx-markdown": "^12.0.1",
-    "ngx-page-scroll": "^7.0.6",
-    "ngx-page-scroll-core": "^7.0.6",
-    "ngx-toastr": "^12.1.0",
-    "npm": "^5.7.1",
+    "ngx-google-analytics": "^14.0.1",
+    "ngx-markdown": "^18.0.0",
+    "ngx-page-scroll": "^13.0.0",
+    "ngx-page-scroll-core": "^13.0.0",
+    "ngx-toastr": "^18.0.0",
     "react": "^16.14.0",
-    "rxjs": "^6.6.7",
-    "rxjs-compat": "^6.6.7",
+    "rxjs": "^7.4.0",
     "ts-helpers": "^1.1.2",
     "ts-md5": "^1.2.4",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.12",
-    "@angular/cli": "^12.2.12",
-    "@angular/compiler-cli": "^12.2.12",
-    "@angular/language-service": "^12.2.12",
+    "@angular-devkit/build-angular": "^18.1.4",
+    "@angular/cli": "^18.1.4",
+    "@angular/compiler-cli": "^18.1.4",
+    "@angular/language-service": "^18.1.4",
     "@types/complexviewer": "^2.1.0",
+    "@types/estree": "^1.0.5",
     "@types/foundation": "~5.5.30",
     "@types/jasmine": "~3.6.0",
     "@types/jquery": "^3.3.36",
-    "@types/node": "~14.0.1",
+    "@types/node": "^20.14.8",
+    "@types/ws": "~8.5.12",
     "codelyzer": "^6.0.0",
     "jasmine-core": "^5.2.0",
     "jasmine-spec-reporter": "^7.0.0",
@@ -72,9 +70,10 @@
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-jasmine": "^5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
+    "npm": "^7.24.2",
     "protractor": "~7.0.0",
     "ts-node": "^5.0.1",
     "tslint": "~6.1.0",
-    "typescript": "~4.3.5"
+    "typescript": "~5.4.5"
   }
 }
